requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
supabase>=2.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# 认证
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# AI客户端
openai==1.10.0
httpx>=0.24,<0.26

# LangChain & LangGraph
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20

# 工具
python-dotenv==1.0.0
redis==5.0.1

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3

# Socket.IO
python-socketio==5.10.0
